LPS-196211 Move logic from SystemObjectDefinitionManagementChecker to ObjectDefinitionUtil
