+++ modules/apps/object/object-api/bnd.bnd
@@ -9,6 +9,7 @@ Export-Package:\
 	com.liferay.object.constants,\
 	com.liferay.object.definition.notification.term.util,\
 	com.liferay.object.definition.tree,\
+	com.liferay.object.definition.util,\
 	com.liferay.object.deployer,\
 	com.liferay.object.display.context,\
 	com.liferay.object.dynamic.data.mapping.expression,\
@@ -34,7 +35,6 @@ Export-Package:\
 	com.liferay.object.service,\
 	com.liferay.object.service.persistence,\
 	com.liferay.object.system,\
-	com.liferay.object.system.util,\
 	com.liferay.object.util.comparator,\
 	com.liferay.object.validation.rule,\
 	com.liferay.object.validation.rule.setting.builder
