+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/definition/util/packageinfo
